+++ packages/eyes-selenium/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - remove type definitions
+- updated to @applitools/eyes-sdk-core@11.4.1 (from 11.3.4)
+- updated to @applitools/visual-grid-client@14.5.15 (from 14.5.7)
 
 ## 4.42.0 - 2020/7/7